Update spdlog to v1.14.1 to fix compiler error on vs2022.

+++ src/device/component/firmwareupdater/FirmwareUpdater.cpp
@@ -28,7 +28,7 @@ FirmwareUpdater::~FirmwareUpdater() noexcept {}
 void FirmwareUpdater::onDeviceFwUpdateCallback(ob_fw_update_state state, const char *message, uint8_t percent, void *userData) {
     FirmwareUpdater *updater = reinterpret_cast<FirmwareUpdater *>(userData);
     if(updater && updater->deviceFwUpdateCallback_) {
-        LOG_DEBUG("Firmware update callback: state={}, message={}, percent={}", state, message, percent);
+        LOG_DEBUG("Firmware update callback: state={}, message={}, percent={}", static_cast<int>(state), message, percent);
         updater->deviceFwUpdateCallback_(state, message, percent);
     }
 }
